+++ examples/minimal-sinatra/app.rb
@@ -1,10 +1,6 @@
 # frozen_string_literal: true
-require 'sinatra/cloudflare_workers'
 require 'sinatra'
 
-get '/' do
-  content_type 'text/plain; charset=utf-8'
-  'minimal sinatra on cloudflare workers'
+get '/frank-says' do
+  'Put this in your pipe & smoke it!'
 end
-
-run Sinatra::Application